Rename files to conform to project naming convention
